Create API.md (#13)
API for test sequence.

Add hardware test commands, and reverse the encoding of 'success' and 'ready' values for 'cnfStart', 'cnfTest' and 'indState' commands so that '0' is the confirmation values and every other value is an error value. The non-zero values are not specified in the API but are potentially useful for debugging.